Refactor/#7 rewrite customfunction implementation errors and magicnumbers (#8)
* refactor: rewrite customfunction implementation errors and-magicnumbers

Replace magic numbers with constants
Protect thread pools with mutexes
Also improve memory safety by replacing int types with size_t types, etc.

* build: UTF8BOM remove

+++ src/image_processor.h
@@ -46,7 +46,7 @@ struct ImageData {
     
     bool is_valid() const {
         return width > 0 && height > 0 && channels > 0 && 
-               pixels.size() == static_cast<size_t>(width * height * channels);
+               pixels.size() == static_cast<size_t>(width) * static_cast<size_t>(height) * static_cast<size_t>(channels);
     }
 };
